# parquetquery iterator less pooling