--- a/tempodb/encoding/vparquet3/block_traceql.go
+++ b/tempodb/encoding/vparquet3/block_traceql.go
@@ -634,12 +634,14 @@ type bridgeIterator struct {
 	cb   traceql.SecondPassFn
 
 	nextSpans []*span
+	at        *parquetquery.IteratorResult
 }
 
 func newBridgeIterator(iter parquetquery.Iterator, cb traceql.SecondPassFn) *bridgeIterator {
 	return &bridgeIterator{
 		iter: iter,
 		cb:   cb,
+		at:   parquetquery.DefaultPool.Get(),
 	}
 }
 
@@ -652,7 +654,7 @@ func (i *bridgeIterator) Next() (*parquetquery.IteratorResult, error) {
 	if len(i.nextSpans) > 0 {
 		ret := i.nextSpans[0]
 		i.nextSpans = i.nextSpans[1:]
-		return spanToIteratorResult(ret), nil
+		return i.spanToIteratorResult(ret), nil
 	}
 
 	for {
@@ -710,13 +712,14 @@ func (i *bridgeIterator) Next() (*parquetquery.IteratorResult, error) {
 		if len(i.nextSpans) > 0 {
 			ret := i.nextSpans[0]
 			i.nextSpans = i.nextSpans[1:]
-			return spanToIteratorResult(ret), nil
+			return i.spanToIteratorResult(ret), nil
 		}
 	}
 }
 
-func spanToIteratorResult(s *span) *parquetquery.IteratorResult {
-	res := parquetquery.DefaultPool.Get()
+func (i *bridgeIterator) spanToIteratorResult(s *span) *parquetquery.IteratorResult {
+	res := i.at
+	res.Reset()
 	res.RowNumber = s.rowNum
 	res.AppendOtherValue(otherEntrySpanKey, s)
 
@@ -735,6 +738,7 @@ func (i *bridgeIterator) SeekTo(to parquetquery.RowNumber, definitionLevel int) 
 
 func (i *bridgeIterator) Close() {
 	i.iter.Close()
+	parquetquery.DefaultPool.Release(i.at)
 }
 
 // confirm rebatchIterator implements parquetquery.Iterator
@@ -743,14 +747,15 @@ var _ parquetquery.Iterator = (*rebatchIterator)(nil)
 // rebatchIterator either passes spansets through directly OR rebatches them based on metadata
 // in OtherEntries
 type rebatchIterator struct {
-	iter parquetquery.Iterator
-
+	iter      parquetquery.Iterator
+	at        *parquetquery.IteratorResult
 	nextSpans []*span
 }
 
 func newRebatchIterator(iter parquetquery.Iterator) *rebatchIterator {
 	return &rebatchIterator{
 		iter: iter,
+		at:   parquetquery.DefaultPool.Get(),
 	}
 }
 
@@ -838,7 +843,8 @@ func (i *rebatchIterator) resultFromNextSpans() *parquetquery.IteratorResult {
 		i.nextSpans = i.nextSpans[1:]
 
 		if ret.cbSpansetFinal && ret.cbSpanset != nil {
-			res := parquetquery.DefaultPool.Get()
+			res := i.at
+			res.Reset()
 			res.AppendOtherValue(otherEntrySpansetKey, ret.cbSpanset)
 			return res
 		}
@@ -853,6 +859,7 @@ func (i *rebatchIterator) SeekTo(to parquetquery.RowNumber, definitionLevel int)
 
 func (i *rebatchIterator) Close() {
 	i.iter.Close()
+	parquetquery.DefaultPool.Release(i.at)
 }
 
 // spansetIterator turns the parquet iterator into the final